Additional fixes for SNR
After watching the SNR feedback for a bit, I was able to
see that it was taking too long for Naomi to realize that
I had stopped talking. I had a fundamental misunderstanding
of what the timeout variable was. I thought it was the number
of frames that Naomi should keep at padding at the beginning
and end, but it turned out to be number of seconds of padding.
10 seconds is a very long time to wait for Naomi to respond.

At this point, SNR is working pretty well.