Asynchronous Say
This Pull Request changes say from synchronous to asynchronous
(sort of). Naomi can now continue to listen while talking, making
it possible to tell Naomi to stop doing what it is doing before
it has finished the current statement. This can be especially
helpful when you get a longer response than you asked for, say
when you ask for weather and Naomi pulls up the news instead,
or if you only want tomorrows weather, but Naomi is going to give
you the whole week's worth anyway.

If you ask Naomi for several things while it is talking, it will
respond to each request in turn as it finishes. This is currently
slightly annoying as it is somewhat difficult to tell which
request Naomi is currently responding to.

naomi/audioengine.py - added a check to the play_fp method to check if
  a "stop" flag has been set. If so, clear the buffer and stop what you
  are doing.

naomi/mic.py - added a "stop" method and an asynchronous "say_thread"
  method. Modified the "say" method to launch the "say_thread" method.

plugins/speechhandler/shutdownplugin/shutdown.py - removed "SHUT" and
  "DOWN" as trigger words. Now this plugin only responds to "SHUTDOWN"
  or "SHUT DOWN". This is so Naomi can respond appropriately to "SHUT
  UP".

plugins/speechhandler/stop - This is a new speechhandler that just
  forces Naomi to stop speaking immediately and clear the queue of
  things waiting to be said when you say "Naomi, stop"