Codacy complaint
Codacy didn't like that I was attempting to access an attribute
of an object (mic) before creating it, even though I was
checking to make sure it existed before accessing it. The idea
was to check if the attribute exists, if so then use it to check
if the process had terminated. If so, start a new process,
otherwise just return. So once again, Codacy has made my code
more complex and difficult to follow by insisting that I fix an
issue that isn't an issue. Thanks Codacy. Good job.