Fixed vad visualization when snr goes negative
I didn't realize that the Signal to Noise Ratio can go
negative on some sound cards. This was causing the display
to grow to more than the size of the terminal window,
resulting in the line wrapping and scrolling. This would
also cause the median and threshold markers to appear in
the wrong place. Also changed the print(phrase) to
the SNR visualization friendly println(phrase).

Fixed the mic.play_file() method to work with the
asnychronous say_thread method, otherwise if Naomi tried
to play a beep or boop to indicate listening but was
already speaking, an error would occur (at least when using
the alsa audio engine). Added visual beep and boop markers
to active_listen when "--print-transcript" option is enabled
to indicate "beep_hi.wav" and "beep_lo.wav".

Added a wait loop to shutdown.py so Naomi can finish
talking before actually shutting down. I feel like I've
done this before, but somehow it doesn't seem to be in
the latest git source.