Intent Parsers
This pull request fundamentally changes the way that Naomi
figures out what you are asking for by introducing three new
intent parsers and the intent parser plugin category, so
additional parsers can be added.

This allows the user to select to use either the Adapt TTI,
Naomi TTI, or Padatious TTI intent parser. It does this by
absorbing intents in a particular format, then parsing them
out in a way that they can be used by those three engines,
then takes the responses and parses them into a standard
format.

To select which intent parser you want to use, set the
"tti_engine" setting in profile.yml to either "Adapt TTI",
"Naomi TTI" or "Padatious TTI".

Naomi TTI is the default and is written in pure Python, so
no additional requirements are needed (you may find that you
need to install the jiwer module if you have not used pip3
with python_requirements.txt recently).

Adapt TTI and Padatious TTI require that you install the
adapt-parser and padatious modules respectively from pypi.
Padatious additionally requires the FANN2 (Fast Artificial
Neural Network) development libraries, and so you might
just want to apt install python3-fann2 if you are lucky
enough to be using a debian derivative.

I ended up touching a lot of files, mostly to get rid of
the need to push copies of the config file from module to
module. Now everything works directly from the profile
module in naomi/profile.py.

I also added a few more plugin settings, but that was too
time consuming for now, so I'll finish that project at
some later time. It is really a separate project.

To add an intent to your plugin, you should create an
intents method. See any speechhandler for an example.

Intents are divided into four basic sections:
keywords - logically grouped sets of words (days, teams,...)
templates - example sentences that should trigger your intent
regex - this is really only used for Adapt, but was so central
  to its function that I couldn't eliminate it. I may try to
  use them in the Naomi TTI also.
action - this is a reference to a function to be called when
  the intent is triggered.

I will post to the support forums about this soon, then try
to update the speechhandler plugin documentation.