Kaldi settings fix
When adding the settings for Kaldi, I missed a couple of libraries.

I want Naomi to be able to translate the title and description
strings from settings, so I enclose them in self.gettext(). This
has to be done before the call back to the superclass, because
that is where the code lives that checks for settings attribute,
unfortunately that is also where the i18n module is used to get
the gettext() method, so it's a bit of a chicken and egg problem.

I used to think that I could just declare the settings as a static
variable, but then that doesn't really work with gettext. At the
moment, I am just creating both gettext and logger before creating
the settings variable, then calling the superclass method right
after. I wonder if I could call the superclass twice?