Fixed lower case keywords
In testing the Naomi TTI I always entered keywords in upper case.
Turns out there is an issue when keywords are entered in lower
case as the input text is converted to upper case before matching.

This converts everything to upper case while parsing the intent.

It also restores the "unclear" functionality, although the unclear
speechhandler no longer functions, since it depended on having the
lowest priority and being valid for everything. Now this is handled
by conversation when it can't figure out which intent to hand off
to.