Moved settings to a method
Moved the settings structure to a method that returns it. This
solved an aweful lot of complexity having to do with calling
super methods in init, and also needing to use self and
gettext while defining the contents of that structure.

Now it is not usually necessary to override the init method

Also added a modification to work with @jeliebig's wakeword
list when returning plugin phrases.

Finally, added gettext to the GenericPlugin init method,
since now all types of plugins need gettext, not just
speechhandler plugins.