Fixed Flite say
Re-added the voice parameter to the say method in the flite
plugin. Removed because codacy was complaining that the declaration
did not match the abstract method it was overriding, but in this
case it is important when testing voices in populate. I should make
this parameter standard for TTS plugins.