Finished adding core variables as settings (#222)
* Finished adding core variables as settings

This depricates the populate.py file for right now. I'm leaving
it in place in case we need to bring it back. For the moment,
all of the settings required for core naomi are contained in
a structure in naomi/application.py. I haven't restored all
the functionality, in particular it would be nice to be
able to restore the ability to preview voices which we had,
at least for flite.

It would be nice to move the structure somewhere more accessible,
like the top of the file or, preferably, into its own file.

I also moved all the methods for sending and reading emails,
anything that requires access to the user's email address,
password, phone number, etc, to the app_utils folder. We may
have gone a bit overboard on the encryption, especially since
plugins still have full access to the user's email, but
it is a nice way of marking that these are variables that
should not be avialable to plugins. I would like to restrict
access to encrypted variables to only the app_utils.py module.

* Codacy complaints