Notifier
This PR re-engages the notifier functionality, and expands the
handle_email_notifications method.

Now when Naomi checks for email, it checks the text of the email
to see if it contains its keyword. If so, then it treats the
email as a request and attempts to handle it and then respond
to the incoming email. Otherwise it just announces that an email
has arrived.

This means that if you send an email or text to Naomi with a
request, you will get an email or text back with a response.

This also introduces