Notifier (#227)
* Finished adding core variables as settings

This depricates the populate.py file for right now. I'm leaving
it in place in case we need to bring it back. For the moment,
all of the settings required for core naomi are contained in
a structure in naomi/application.py. I haven't restored all
the functionality, in particular it would be nice to be
able to restore the ability to preview voices which we had,
at least for flite.

It would be nice to move the structure somewhere more accessible,
like the top of the file or, preferably, into its own file.

I also moved all the methods for sending and reading emails,
anything that requires access to the user's email address,
password, phone number, etc, to the app_utils folder. We may
have gone a bit overboard on the encryption, especially since
plugins still have full access to the user's email, but
it is a nice way of marking that these are variables that
should not be avialable to plugins. I would like to restrict
access to encrypted variables to only the app_utils.py module.

* Codacy complaints

* Notifier

This PR re-engages the notifier functionality, and expands the
handle_email_notifications method.

Now when Naomi checks for email, it checks the text of the email
to see if it contains its keyword. If so, then it treats the
email as a request and attempts to handle it and then respond
to the incoming email. Otherwise it just announces that an email
has arrived.

This means that if you send an email or text to Naomi with a
request, you will get an email or text back with a response.

This also introduces

* Protect encrypted values

This protects encrypted values by only allowing whitelisted modules
to access them. This means that a plugin would not be able to
access your email address or password, for instance.

* Removed debugging statements

* Codacy cleanup