Rebuilt translations
I re-built the translation files. There were so many changes
in 2019 with the new template system for building intents and
the settings built into the plugins directly rather than in
the populate program, that I decided to just start over and
regenerate the translation files from scratch.

I also changed the underlying engine for generating and updating
translation files from pygettext.py (which keeps changing location)
to xgettext, which only requires installing the gettext apt
package.

Almost every set of translations now have some lines that I had
to generate with the help of google translate, so they all need
to be checked for accuracy.