Email bug (#236)
* Email bug

When running naomi, the check_imap_config() and check_smtp_config()
functions in naomi/app_utils.py have a bug in them which causes
Naomi to crash if email is not set up.

* Codacy fix

Apparently codacy uses an older version of python which does not
include the ConnectionRefusedError exception, so it was choking
on that line. I have changed the exception to ConnectionError which
ConnectionRefusedError is a subclass of, and then logged the
details of the exception.

* More fun with Codacy

I can't figure out what version codacy is using, so trying to force
it to use 3, which should have the ConnectionError exception.

* More Codacy

Codacy is apparently using a python 3 version below 3.3 when
ConnectionError and ConnectionRefusedError were added.
Apparently it used to use OSError instead. I tried using that
with Python 3.8 and that seems to capture the error, so I'm
trying that.

* More Codacy

Added requests because one website suggested that ConnectionError
was defined in requests. Figured out that requests.ConnectionError
was probably different than the built-in ConnectionError, I am
getting, but accidentally left the import statement in there.