General debugging updates
These are general updates that seem to help Naomi run better for
me. The biggest updates are having the pocketsphinx plugin
automatically download and configure a language model, and a fix
that caused mpdcontrol to fail "I'm sorry, I had some trouble
with that operation..." if keyword is a list rather than a single
value.

Other than that, there are some tweaks to the intent parser. The
word "WHAT" in a request was strongly linked to the LIFE plugin
when using the naomi tti program, so I have reduced the weight
of a word by the number of intents that word appears in, causeing
Naomi to respond with "Life, don't talk to me about life" when
asked "Naomi, what is the weather". I also was having trouble
getting news to run instead of hacker news, so I added the word
"news" to every intent template in the News plugin.