General Debugging (#239)
* Added process_completedprocess function

Added process_completedprocess(completedprocess, output) function
to naomi/run_command.py. This transforms a completedprocess object
into a success or error message. This will be used when setting up
a new pocketsphinx installation and for STT training.

* General debugging updates

These are general updates that seem to help Naomi run better for
me. The biggest updates are having the pocketsphinx plugin
automatically download and configure a language model, and a fix
that caused mpdcontrol to fail "I'm sorry, I had some trouble
with that operation..." if keyword is a list rather than a single
value.

Other than that, there are some tweaks to the intent parser. The
word "WHAT" in a request was strongly linked to the LIFE plugin
when using the naomi tti program, so I have reduced the weight
of a word by the number of intents that word appears in, causeing
Naomi to respond with "Life, don't talk to me about life" when
asked "Naomi, what is the weather". I also was having trouble
getting news to run instead of hacker news, so I added the word
"news" to every intent template in the News plugin.

* Move UTF-8 encoding

Instead of having to manually encode each line to UTF-8 as it is
added, which is difficult and error prone, just add each line as
a string, then convert the whole thing to UTF-8 as the response
is about to be returned to the web server.

* Intent Verification

In this commit, I am adding the ability to see and verify the
intent that Naomi selected to respond to your query.

I also have made a number of minor bug fixes as I am testing.

NaomiSTTTrainer.py - added the ability to view and verify the intent
selected by Naomi. This isn't currently used for training purposes,
but I envision adding templates based on what the user actually says
and possibly also modifying the intent weights.

naomi/__main__.py - added an option allowing the user to set Naomi
to "listen while talking". This is because I have a couple of
microphone hats that cause Naomi to react to its own voice. When
using headphones, this is not a problem. It also seems to be less
of an issue when using an integrated USB sound card like a USB
conference phone. But for right now, I am making the default to
not listen while talking. I have had a couple of thoughts, such as
only reacting to the keyword "stop" when talking.

naomi/app_utils.py - wrapped the "is_negative()" and "is_positive()"
regex in gettext so different filters can be set for different
languages.

naomi/application.py - use profile.set_arg() to store the state of
several settings, so I have access from any module without manually
passing the values through.

naomi/brain.py - store intent and score while processing a request.

naomi/mic.py - say() can be synchronous or asynchronous depending
on whether you want Naomi to listen while talking.

plugins/speechhandler/life/life.py - fix an issue with the Fourty
Two answer (return string instead of list)

plugins/speechhandler/shutdownplugin/shutdown.py - Clean up
correctly before exiting whether or not "mic.say()" is synchronous.

plugins/tti/naomi_tti/naomi_tti.py - added a list of words to ignore
when weighting intents. Cleaned up some debugging output.

plugins/vad/snr_vad/snr_vad.py - Changed the threshold for stopping
recording to 1/2 the distance betweeen mean and threshold. This was
because I notice that I usually start talking loudly, but then lose
volume towards the end of my request, causing the end to get cut
off. I think this is because my Naomi has an LED light that indicates
when it is listening, and when I can see that I have Naomi's
attention I stop speaking so forcefully. This sometimes resulted
in the last few words of a request not being heard.

* Codacy complaints

Fixed a few trailing whitespaces and naked excepts.

* Skipped plugin reporting

Report skipped speechhandler plugins, but don't report other types
of plugins that are probably being skipped because they just
haven't been set up.