Separate intents by locale
It has occured to me that defining intents in english and then
using gettext to translate the strings into other languages will
not always work. Unlike simply translating an output string to
another language, it is not guaranteed that there are the same
number of common ways of saying something in every language, or
even that the response will be structured in terms of how keywords
are defined.

I have reorganized the intents to store keywords, regex's and
templates under different locale's. If the user's selected
locale does not exist, then I use the two letter language code
to select the first locale using the same language.