Better recovery after canceling setup
I started setting up Naomi, but canceled the process before it
completed. At that point, it had built the Pocketsphinx model,
but had not set the locations of the hmm dir or fst file in the
profile. Because it did not have to create them the next time I
ran it, it did not set the default location of the fst file.