naomi-setup.sh
Got the naomi-setup.sh script working with the virtualenvwrapper
option.

Discovered that the modules installed for the main python instance
do not automatically become available for the virtual python
environment of the same type, so using apt to install
python3-pocketsphinx did not work, since the pocketsphinx module
would not be available to the virtual environment.

I could have just installed the Pypi version through pip, which
should be fine, but it could lead to a situation where the
python module is a different version from the main pocketsphinx.

It's not that big a deal to build pocketsphinx, so I'm opting to
just do that and install the python version that matches.