Install with custom python
This fixes the issues with installing Naomi with a custom copy of
python in order to protect the system Python from the Naomi
package requirements.

I had to upgrade the version of python since the latest version of
mock required python3.6, but we were using 3.5. That meant having
to also update the verification code and signature used for checking
the signature hash. Newer versions of python appear to automatically
build their own setuptools and pip to keep everything synchronized
so I removed the sections of code that did that manually.

I moved all the source code used to build python into the
~/.config/naomi/sources folder which can be safely deleted after
building Naomi. The pip cache should probably also be located
under this folder, but for the moment I have left it in
~/.config/naomi/local/cache.