Check for git folders before cloning
If the folder already exists, skip the cloning. I have a bad internet connection here and keep getting network errors in the middle of cloning or downloading source code. I put a check to make sure the clone command worked before moving on. Unfortunately, this meant that if I got the "The directory exists and is not empty" error, the program would just stop. So I check to see if the directory exists before attempting the clone command. This works as long as the folder does not exist (which it shouldn't).