Fix NPE lists and extra email fields
This fixes two problems. First, for some reason the last time
I tried to install a plugin with a space in the name, like
"You're Welcome", it broke into two parameters so Naomi tried
to install a plugin called "You're" and a plugin called "Welcome"
even though the name was surrounded with quotes.

I have fixed this by no longer allowing lists of plugin names.
Now instead of:
./Naomi --install "You're Welcome" "Control LED"
you have to use:
./Naomi --install "You're Welcome" --install "Control LED"
Since most people will only be installing one plugin at a time,
I don't think that's a big issue. I also had to change the
code for flattening the associated array.

The other problem was that even if you did not enter an email
address, Naomi would continue to ask you email related questions.
This was because we changed the ["email", "address"] profile
setting into a password. If you retrieve it like a normal
setting, then you will get a hash even if the value itself is
the empty string.