New NotificationClient plugin type
Notifications can now be written as plugins. This demonstrates
writing a NotificationClient plugin that checks an email address
for new emails. If the email contains one of Naomi's wake words,
the email is parsed as a request and Naomi's response is returned
to the emailer. Otherwise, Naomi announces the email to the user.

This is a separate thread that polls every 30 seconds and can
respond to an external change.

Right now, there is only one "notifier" thread that contains
all of the "notificationclient" objects, so you can't set one
notification to run every 30 seconds and another to run every
12 hours, for instance. The 30 seconds is also hard coded and
should be replaced with a profile setting that the user could
override.

If a user wants to create a notification object, they just need
to create an object that inherits from NotificationClientPlugin
and has a "gather" method (as demonstrated by the respondtoemail
plugin, you can prevent a plugin from operating simply by not
setting the "gather" method to anything). The "gather()" method
will be called every time the notifier runs.

This is a fairly complex example, and has some privacy issues,
since right now it doesn't filter the "from" email and will
respond to any email or text message that contains its wake word.
We should probably include a simpler example and move this to
the exchange.