Don't respond to emails by default
The respond to emails feature seems a little dangerous because it
currently only requires the user to set up email. Then any email
recieved with the wake word will cause Naomi to respond as if the
email is a command, sending back Naomi's response. For the most
part this is fine with the current News, weather, etc. but could
be used to either knowingly or accidentally get information that
the user might not want put out there.

For instance, if Naomi is connected to your google calendar, and
you get an email from your boss saying "Interesting to hear about
Naomi yesterday. Do you have tomorrow's meeting on your calendar"
Naomi might see "Naomi", "tomorrow" and "calendar" and respond with
"You have the following events scheduled on your calendar: 7:30 am
call in sick, 8:00 am meet friends, 9am hike to waterfall." As
Naomi becomes more connected and learns more about you, this would
become more of an issue.

This change makes it so that the user has to specifically enable
the "respond" functionality by creating a "respond_to_emails"
entry in profile.yml. This can contain a whitelist of email
addresses that Naomi is allowed to respond to.

The following rules apply:
No entry - Do not respond to emails. Simply announce when new emails
arrive.

Empty list - respond to any email containing the wake word.
```
respond_to_emails: []
```

List of addresses - respond to any email from the list of email
"from" addresses that contains a wake word:
```
respond_to_emails:
- aaronchantrill@projectnaomi.com
- 2125551212@vtext.com
```