Use bombocher/pocketsphinx-python
The version of pocketsphinx-python in cmusphinx/pocketsphinx-python
is a fork of the bombocher repository. Both repositories have
diverged, but bombocher reports a larger version number (0.1.15
as compared to 0.0.9 for the cmusphinx version) and is the source
for the pip module, and has a lot of features that would be really
useful to have access to.