Pocketsphinx python (#278)
* Use bambocher/pocketsphinx-python

The version of pocketsphinx-python in cmusphinx/pocketsphinx-python
is a fork of the bombocher repository. Both repositories have
diverged, but bombocher reports a larger version number (0.1.15
as compared to 0.0.9 for the cmusphinx version) and is the source
for the pip module, and has a lot of features that would be really
useful to have access to.

* Update apt_requirements.txt

Bambocher's version of pocketsphinx-python seems to require libpulse-dev to install. Since we want to use pulse anyway, this should be fine. I added libpulse-dev to the list of required apt packages under pocketsphinx, because otherwise you get this error message when building:
```
deps/sphinxbase/src/libsphinxad/ad_pulse.c:44:10: fatal error: pulse/pulseaudio.h: No such file or directory
```