Update Mozilla DeepSpeech (#300)
* Update Mozilla DeepSpeech

Update the implementation of Mozilla's DeepSpeech STT engine.
This version tries to set up deepspeech as much as possible. It
downloads the source code to ~/.config/naomi/sources/STTv{version}
and the acoustic model to ~/.config/naomi/deepspeech/model_{version}.pbmm

The big improvement here is that it uses the KenLM engine to build
language models based on phrases that Naomi is likely to hear based
on the templates provided in the speechhandler plugins. KenLM pays
attention to words and word order, so it will be interesting to see
how word error rate compares to pocketsphinx which just listens for
individual words from a dictionary, or how word error rates improve
if we can get pocketsphinx to use kenlm also.

I had to modify all the text to intent parsers to return a list of
full phrases from the intents instead of a list of words. This seems
to work fine with pocketsphinx as it is, but allows KenLM to
analyze word order.

I had to make one small change to sphinxvocab, where I do not
include keywords surrounded by curly brackets when passing the list
of words to Phonetisaurus.

* Minor fixes for Deepspeech

This adds the requirement for the deepspeech python module plus
the apt requirements for Naomi to set up the tools needed to
generate language models. It also now correctly downloads the
tensorflow-lite version of the language model when running on
the Raspberry Pi.

* Provide additional details

Provide some additional details about where the DeepSpeech source
code is being downloaded to.

* Update sphinxvocab.py

Removed an errant print