Fixed deepspeech model download
There were a couple of issues here. One was that if Naomi needed
to download the model, it reconstructed the URL without paying
attention to the architecture, so on the raspberry pi it still
downloaded the pbmm model instead of the tflite model. I also
seemed to be having some trouble downloading the binary correctly
and kept getting an error message about the file header being
malformed. I "borrowed" some code from stack overflow which seems
to work pretty well. I put this code into app_utils as a start to
working on Issue #303 - Standard functions with progress bars for
downloads. At least we don't have to reinvent the wheel every
time we need to download a binary file.