Fix audiodevice output buffer underrun error (#309)
* Change Naomi TTI score to a confidence percent

Just a slight change to the Naomi TTI scoring. After generating
scores for all the intents, add those scores up and divide
individual scores by the total. This makes the score always fall
between 0 and 1 (currently there is no upper limit and I have
seen scores as high as between 6 and 7).

This should also make Naomi more confident in some cases. Right
now if one intent scores .01 and all the other intents are zero,
Naomi will respond with "unclear" since no intent has scored
above .05 (5%). With this new system, .01/.01=1 (100%). I don't
know if this will turn out to be a good thing or not. I think
it will be good, but if it causes problems, we might need to add
a filter before the conversion.

* committing change so I can pull upstream

* Fix buffer underrun error

After the latest Raspbian Buster update on my Raspberry Pi 4, I
was getting the following error message:

ALSA lib pcm.c:8424:(snd_pcm_recover) underrun occurred

This was causing Naomi to make raspberry noises instead of
talking when trying to play audio generated from flite.

I tried increasing the chunk size in the profile.yml file, but
it did not seem to affect the playback of a file in play_fp. It
looks like the chunksize and output padding both get written to
attributes of the audioengine plugin but I didn't see anywhere
that they were actually passed to the play_file or play_fp
audiodevice methods. Since we can read the profile directly from
any method just by including the profile module, I'm just reading
them directly from the play_fp method.

Increasing the chunk size worked, but the audio kept getting
clipped at the end. I tried enabling the padding option in
profile.yml, but that didn't help. I ended up adding a new
option under  audio -> output_pause to have the computer pause
after the last chunk is written to the device and before the
stream is closed, and that finally worked. I set it to 0.5
seconds on my system which seemed to work fine. 0.1 was not
enough to prevent clipping.