Fix play_fp parameters
In pull request #309, I removed the chunksize and add_padding
optional parameters from play_fp, and pulled those values
directly from the profile.

This worked fine in normal operation, but introduced issues with
the configuration routines in application.py which are passing
parameters to play_file which are being passed along to play_fp.

This update restores legacy compatibility by using *args and
**kwargs to allow values to be passed in optionally, which makes
it compatible with the call from play_file. If chunksize and
add_padding are passed in as keyword arguments for some reason,
those values are used instead of the profile values. In general,
developers should not pass values directly to play_file or play_fp.

I also removed the parameters from the play_file calls in
application.py.