Expect and Confirm
This demonstrates the use of the Expect and Confirm functions.
This is a first step in a demonstration project for a new
conversation engine.

Right now, this only works when the "listen when speaking" option
is turned off. I'll work this week on getting these functions to
work with the asyncronous speaking.

The syntax is somewhat convoluted. The return value is a tuple
consisting of both a boolean indicating whether the computer
thinks it heard one of the options and a second parameter that
either returns the option it believes it heard (so the author
can do exact matching against the list of options passed in) or
the transcript generated by the active listener. At the moment,
the conversation engine does not check for the wake word when
handling this active listener transcription. This is so that
"conversation" can examine the output of your plugin and handle
the transcription if the plugin was unable to.

I have included a new implementation of the "joke" speechhandler
which demonstrates the use of both new functions, plus provides a
proper return value to "conversation".

When using either, use "(matched, phrase) = mic.expect(name, phrases)"
or "(matched, phrase) = mic.confirm(prompt)". Check the "matched"
parameter to see if any of the phrases passed in were matched. If
this is True, then you can go ahead and match against the list of
options passed in. If this is false, then the user was most likely
not responding to your prompt, and the active listener transcription
should be passed back to "conversation" for intent parsing and
handling. I know that's a really confusing word salad, but please
examine the "joke" plugin for examples.

You can try this out with regular Naomi or with the --local mic.
A local conversation looks something like this:

COMPUTER: My name is COMPUTER.
COMPUTER: How can I be of service?
YOU: tell me a joke
JokeIntent 0.8555247923483513
COMPUTER: Knock knock
YOU: who's there?
COMPUTER: arizona
YOU: arizona who?
COMPUTER: arizona room for only one of us in this room
COMPUTER: Was that joke funny?
YOU: what time is it?
ClockIntent 1.0
COMPUTER: It is  8:44 PM right now.

As you can see, I was able to escape from the confirmation dialog
by entering a query that activated another intent without having
to answer the question.

This has a couple of problems. First, if you are using async_say
(listen while talking) it will start listening for the response
before it asks the question, and may answer its own question if
it hears itself talking. I'm going to try to add this to the
"say" queue so it will run in a synchronous way with "say".
Second, it needs some cruft added to the dictionary because right
now it hears almost everything as its expected words regardless of
what you actually say. Third, I really should use the user's
selected intent parser to match what naomi hears to the phrases
but right now I'm just using edit distance, and a threshold of
50% which is hard coded in.