Handle different return types
I forgot that the return type from an intent is supposed to be a
tuple containing whether or not the intent was able to handle the
request along with a transcription to be fed back to brain if not.

I was just returning true from plugins to indicate that the
operation had succeeded, which conversation did not like. So
now I am handling any form of output from the plugins.