Uninstall
${NAOMI_BASE}/naomi-setup.sh was not passing arguments to
${NAOMI_BASE}/installers/naomi-setup.sh, so passing arguments
to the first script had no effect on the second, which was
expecting the arguments.

I also changed the argument from "-uninstall" or "uninstall" to
"--uninstall" which is a more typical way to pass arguments in
Linux. I also changed the "Continue?" prompt to accept lower
case y/n and used SUDO_COMMAND to give the user a chance to
cancel if they don't want to delete the ~/Naomi or ~/.config/naomi
directories.