Settings for notificationclient
Added settings for the demo notification client so it now will
prompt the user for a list of email addresses it is allowed to
respond to.

Also changed the "allow_multiple" parameter for settings to
"return_type", which I think will be more flexible. In this case,
you can set the return type to "list" which will convert your
string response into a list of strings. It can also return values
as boolean, number, or str (the default). This currently only
works with the default input control type, although it could
easily be extended to other types.