Remove unused variable
I'm trying to figure out what is the more flexible option - a
return_type parameter, or an accept_list parameter. The
return_type parameter seems more flexible, except that the type
of response is implied in the name of the control (text, boolean,
number, etc). If I set it so that a boolean control actually
returns a boolean or a number returns a number, then I can use
"accept_list" to return a list of booleans or a list of numbers.