Multiple keywords fix (#337)
* Various fixes

This includes a fix for multiple keywords and some small changes to
the NaomiSTTTrainer.py app.

* Removed extra parameter

Removed an extra "true" from a call to simple input for
"allow multiple", a setting I just removed and will deal with at
the setting level instead.

* Settings for notificationclient

Added settings for the demo notification client so it now will
prompt the user for a list of email addresses it is allowed to
respond to.

Also changed the "allow_multiple" parameter for settings to
"return_type", which I think will be more flexible. In this case,
you can set the return type to "list" which will convert your
string response into a list of strings. It can also return values
as boolean, number, or str (the default). This currently only
works with the default input control type, although it could
easily be extended to other types.

* Fix "noise" output

* Remove unused variable

I'm trying to figure out what is the more flexible option - a
return_type parameter, or an accept_list parameter. The
return_type parameter seems more flexible, except that the type
of response is implied in the name of the control (text, boolean,
number, etc). If I set it so that a boolean control actually
returns a boolean or a number returns a number, then I can use
"accept_list" to return a list of booleans or a list of numbers.

* Change back to "return_list"

* Fix greeting

When using multiple keywords, Naomi greets the user with "My name
is <name1> but you can also call me <name2> or <name3>". This
fixed the "or" join so "or" was not appended to the previous name.